+++ offlineimap/imapserver.py
@@ -55,6 +55,8 @@ class IMAPServer:
         self.tunnel = repos.getpreauthtunnel()
         self.usessl = repos.getssl()
         self.username = None if self.tunnel else repos.getuser()
+        self.user_identity = repos.get_remote_identity()
+        self.authmechs = repos.get_auth_mechanisms()
         self.password = None
         self.passworderror = None
         self.goodpassword = None
@@ -133,6 +135,24 @@ class IMAPServer:
         self.ui.debug('imap', 'Attempting IMAP LOGIN authentication')
         imapobj.login(self.username, self.getpassword())
 
+
+    def plainhandler(self, response):
+        """
+        Implements SASL PLAIN authentication, RFC 4616,
+          http://tools.ietf.org/html/rfc4616
+
+        """
+        authc = self.username
+        passwd = self.getpassword()
+        authz = ''
+        if self.user_identity != None:
+            authz = self.user_identity
+        NULL = u'\x00'
+        retval = NULL.join((authz, authc, passwd)).encode('utf-8')
+        self.ui.debug('imap', 'plainhandler: returning %s' % retval)
+        return retval
+
+
     def gssauth(self, response):
         data = base64.b64encode(response)
         try:
@@ -180,7 +200,7 @@ class IMAPServer:
         tried_to_authn = False
 
         # Try GSSAPI and continue if it fails
-        if 'AUTH=GSSAPI' in imapobj.capabilities and have_gss:
+        if 'AUTH=GSSAPI' in imapobj.capabilities and have_gss and "GSSAPI" in self.authmechs:
             self.connectionlock.acquire()
             self.ui.debug('imap', 'Attempting GSSAPI authentication')
             tried_to_authn = True
@@ -213,7 +233,9 @@ class IMAPServer:
                   "TLS connection: %s" % str(e),
                   OfflineImapError.ERROR.REPO)
 
-        if 'AUTH=CRAM-MD5' in imapobj.capabilities:
+    # Hashed authenticators come first: they don't reveal
+    # passwords.
+        if 'AUTH=CRAM-MD5' in imapobj.capabilities and "CRAM-MD5" in self.authmechs:
             tried_to_authn = True
             self.ui.debug('imap', 'Attempting '
               'CRAM-MD5 authentication')
@@ -224,23 +246,36 @@ class IMAPServer:
                 self.ui.warn('CRAM-MD5 authentication failed: %s' % e)
                 exc_stack.append(('CRAM-MD5', e))
 
-        # Last resort: use LOGIN command,
-        # unless LOGINDISABLED is advertized (RFC 2595)
-        if 'LOGINDISABLED' in imapobj.capabilities:
-            e = OfflineImapError("IMAP LOGIN is "
-              "disabled by server.  Need to use SSL?",
-               OfflineImapError.ERROR.REPO)
-            exc_stack.append(('IMAP LOGIN', e))
-        else:
+        # Try plaintext authenticators.
+        if 'AUTH=PLAIN' in imapobj.capabilities and "PLAIN" in self.authmechs:
             tried_to_authn = True
             self.ui.debug('imap', 'Attempting '
-              'IMAP LOGIN authentication')
+              'PLAIN authentication')
             try:
-                self.loginauth(imapobj)
+                imapobj.authenticate('PLAIN', self.plainhandler)
                 return
             except imapobj.error as e:
-                self.ui.warn('IMAP LOGIN authentication failed: %s' % e)
+                self.ui.warn('PLAIN authentication failed: %s' % e)
+                exc_stack.append(('PLAIN', e))
+
+        # Last resort: use LOGIN command,
+        # unless LOGINDISABLED is advertized (RFC 2595)
+        if 'LOGIN' in self.authmechs:
+            if 'LOGINDISABLED' in imapobj.capabilities:
+                e = OfflineImapError("IMAP LOGIN is "
+                  "disabled by server.  Need to use SSL?",
+                   OfflineImapError.ERROR.REPO)
                 exc_stack.append(('IMAP LOGIN', e))
+            else:
+                tried_to_authn = True
+                self.ui.debug('imap', 'Attempting '
+                  'IMAP LOGIN authentication')
+                try:
+                    self.loginauth(imapobj)
+                    return
+                except imapobj.error as e:
+                    self.ui.warn('IMAP LOGIN authentication failed: %s' % e)
+                    exc_stack.append(('IMAP LOGIN', e))
 
         if len(exc_stack):
             msg = "\n\t".join(map(